test: тесты сервисов отделены от тестов свойств моделей
