+++ server/apps/dcis/tests/models/__init__.py
@@ -1,0 +1,1 @@
+from .document import DocumentModelTestCase
