+++ server/apps/dcis/tests/services/__init__.py
@@ -1,8 +1,8 @@
 from .column_dimension_services import ColumnDimensionTestCase
-from .divisions_services import DivisionsServicesTestCase
+from .divisions_services import DivisionTestCase
 from .document_services import DocumentTestCase, GetUserDocumentsTestCase
 from .period_services import GetUserPeriodsTestCase, PeriodTestCase
-from .privilege_services import PrivilegeServicesTestCase
+from .privilege_services import PrivilegeTestCase
 from .project_services import GetUserProjectsTestCase, ProjectTestCase
 from .row_dimension_services import RowDimensionTestCase
 from .status_services import StatusTestCase
